Adds a device metaclass.
